Let the fields handle the widgets attrs
